[Feature #1628] Create UI for exporting glossary with skos:exactMatch references.
