+++ src/action/AsyncVocabularyActions.ts
@@ -1,17 +1,22 @@
 import { ThunkDispatch } from "../util/Types";
+import * as SyncActions from "./SyncActions";
 import {
   asyncActionFailure,
   asyncActionRequest,
+  asyncActionSuccess,
   asyncActionSuccessWithPayload,
   publishMessage,
 } from "./SyncActions";
-import { IRI } from "../util/VocabularyUtils";
+import VocabularyUtils, { IRI } from "../util/VocabularyUtils";
 import ActionType from "./ActionType";
-import Ajax, { param } from "../util/Ajax";
+import Ajax, { param, params } from "../util/Ajax";
 import Constants from "../util/Constants";
 import { ErrorData } from "../model/ErrorInfo";
 import Message from "../model/Message";
 import MessageType from "../model/MessageType";
+import ExportType from "../util/ExportType";
+import { AxiosResponse } from "axios";
+import Utils from "../util/Utils";
 
 export function loadTermCount(vocabularyIri: IRI) {
   const action = { type: ActionType.LOAD_TERM_COUNT, vocabularyIri };
@@ -38,4 +43,57 @@ export function loadTermCount(vocabularyIri: IRI) {
         return dispatch(publishMessage(new Message(error, MessageType.ERROR)));
       });
   };
+}
+
+export function exportGlossary(
+  vocabularyIri: IRI,
+  type: ExportType,
+  queryParams: {} = {}
+) {
+  const action = {
+    type: ActionType.EXPORT_GLOSSARY,
+  };
+  return (dispatch: ThunkDispatch) => {
+    dispatch(asyncActionRequest(action));
+    const url =
+      Constants.API_PREFIX +
+      "/vocabularies/" +
+      vocabularyIri.fragment +
+      "/terms";
+    return Ajax.getRaw(
+      url,
+      params(queryParams)
+        .param("namespace", vocabularyIri.namespace)
+        .accept(type.mimeType)
+        .responseType("arraybuffer")
+    )
+      .then((resp: AxiosResponse) => {
+        const disposition = resp.headers[Constants.Headers.CONTENT_DISPOSITION];
+        const filenameMatch = disposition
+          ? disposition.match(/filename="(.+\..+)"/)
+          : null;
+        if (filenameMatch) {
+          const fileName = filenameMatch[1];
+          Utils.fileDownload(resp.data, fileName, type.mimeType);
+          return dispatch(asyncActionSuccess(action));
+        } else {
+          const error: ErrorData = {
+            requestUrl: url,
+            messageId: "vocabulary.summary.export.error",
+          };
+          dispatch(asyncActionFailure(action, error));
+          return dispatch(
+            SyncActions.publishMessage(new Message(error, MessageType.ERROR))
+          );
+        }
+      })
+      .catch((error: ErrorData) => dispatch(asyncActionFailure(action, error)));
+  };
+}
+
+export function exportGlossaryWithExactMatchReferences(vocabularyIri: IRI) {
+    return exportGlossary(vocabularyIri, ExportType.Turtle, {
+        withReferences: true,
+        property: [VocabularyUtils.SKOS_EXACT_MATCH]
+    });
 }
