+++ src/action/__tests__/AsyncActions.test.ts
@@ -6,7 +6,6 @@ import {
   createVocabulary,
   executeFileTextAnalysis,
   exportFileContent,
-  exportGlossary,
   getContentType,
   getLabel,
   getProperties,
@@ -69,7 +68,6 @@ import TermItState from "../../model/TermItState";
 import Resource, { CONTEXT as RESOURCE_CONTEXT } from "../../model/Resource";
 import Utils from "../../util/Utils";
 import AsyncActionStatus from "../AsyncActionStatus";
-import ExportType from "../../util/ExportType";
 import fileContent from "../../rest-mock/file";
 import TermItFile from "../../model/File";
 import MessageType from "../../model/MessageType";
@@ -1431,148 +1429,6 @@ describe("Async actions", () => {
         expect(Ajax.get).toHaveBeenCalled();
         expect(result).toBeDefined();
         expect(result.length).toEqual(0);
-      });
-    });
-  });
-
-  describe("exportGlossary", () => {
-    it("provides vocabulary normalized name and namespace in request", () => {
-      const namespace =
-        "http://onto.fel.cvut.cz/ontologies/termit/vocabularies/";
-      const name = "test-vocabulary";
-      Ajax.getRaw = jest.fn().mockImplementation(() =>
-        Promise.resolve({
-          data: "test",
-          headers: { "Content-type": "text/csv" },
-        })
-      );
-      return Promise.resolve(
-        (store.dispatch as ThunkDispatch)(
-          exportGlossary(
-            {
-              namespace,
-              fragment: name,
-            },
-            ExportType.CSV
-          )
-        )
-      ).then(() => {
-        expect(Ajax.getRaw).toHaveBeenCalled();
-        const url = (Ajax.getRaw as jest.Mock).mock.calls[0][0];
-        expect(url).toEqual(
-          Constants.API_PREFIX + "/vocabularies/" + name + "/terms"
-        );
-        const config = (Ajax.getRaw as jest.Mock).mock.calls[0][1];
-        expect(config.getParams().namespace).toEqual(namespace);
-      });
-    });
-
-    it("sets accept type to CSV when CSV export type is provided", () => {
-      const iri = VocabularyUtils.create(Generator.generateUri());
-      Ajax.getRaw = jest.fn().mockImplementation(() =>
-        Promise.resolve({
-          data: "test",
-          headers: { "Content-type": "text/csv" },
-        })
-      );
-      return Promise.resolve(
-        (store.dispatch as ThunkDispatch)(exportGlossary(iri, ExportType.CSV))
-      ).then(() => {
-        expect(Ajax.getRaw).toHaveBeenCalled();
-        const config = (Ajax.getRaw as jest.Mock).mock.calls[0][1];
-        expect(config.getHeaders()[Constants.Headers.ACCEPT]).toEqual(
-          Constants.CSV_MIME_TYPE
-        );
-      });
-    });
-
-    it("sets accept type to Excel when Excel export type is provided", () => {
-      const iri = VocabularyUtils.create(Generator.generateUri());
-      Ajax.getRaw = jest.fn().mockImplementation(() =>
-        Promise.resolve({
-          data: "test",
-          headers: { "Content-type": Constants.EXCEL_MIME_TYPE },
-        })
-      );
-      return Promise.resolve(
-        (store.dispatch as ThunkDispatch)(exportGlossary(iri, ExportType.Excel))
-      ).then(() => {
-        expect(Ajax.getRaw).toHaveBeenCalled();
-        const config = (Ajax.getRaw as jest.Mock).mock.calls[0][1];
-        expect(config.getHeaders()[Constants.Headers.ACCEPT]).toEqual(
-          Constants.EXCEL_MIME_TYPE
-        );
-      });
-    });
-
-    it("invokes file save on request success", () => {
-      const iri = VocabularyUtils.create(Generator.generateUri());
-      const data = "test";
-      const fileName = "test.csv";
-      Ajax.getRaw = jest.fn().mockImplementation(() =>
-        Promise.resolve({
-          data,
-          headers: {
-            "content-type": Constants.CSV_MIME_TYPE,
-            "content-disposition": 'attachment; filename="' + fileName + '"',
-          },
-        })
-      );
-      Utils.fileDownload = jest.fn();
-      return Promise.resolve(
-        (store.dispatch as ThunkDispatch)(exportGlossary(iri, ExportType.CSV))
-      ).then(() => {
-        expect(Utils.fileDownload).toHaveBeenCalledWith(
-          data,
-          fileName,
-          Constants.CSV_MIME_TYPE
-        );
-      });
-    });
-
-    it("dispatches async success on request success and correct data", () => {
-      const iri = VocabularyUtils.create(Generator.generateUri());
-      const data = "test";
-      const fileName = "test.csv";
-      Ajax.getRaw = jest.fn().mockImplementation(() =>
-        Promise.resolve({
-          data,
-          headers: {
-            "content-type": Constants.CSV_MIME_TYPE,
-            "content-disposition": 'attachment; filename="' + fileName + '"',
-          },
-        })
-      );
-      Utils.fileDownload = jest.fn();
-      return Promise.resolve(
-        (store.dispatch as ThunkDispatch)(exportGlossary(iri, ExportType.CSV))
-      ).then(() => {
-        expect(store.getActions().length).toEqual(2);
-        const successAction = store.getActions()[1];
-        expect(successAction.type).toEqual(ActionType.EXPORT_GLOSSARY);
-        expect(successAction.status).toEqual(AsyncActionStatus.SUCCESS);
-      });
-    });
-
-    it("dispatches failure when response does not contain correct data", () => {
-      const iri = VocabularyUtils.create(Generator.generateUri());
-      const data = "test";
-      Ajax.getRaw = jest.fn().mockImplementation(() =>
-        Promise.resolve({
-          data,
-          headers: {
-            "content-type": Constants.CSV_MIME_TYPE,
-          },
-        })
-      );
-      Utils.fileDownload = jest.fn();
-      return Promise.resolve(
-        (store.dispatch as ThunkDispatch)(exportGlossary(iri, ExportType.CSV))
-      ).then(() => {
-        expect(store.getActions().length).toEqual(3);
-        const successAction = store.getActions()[1];
-        expect(successAction.type).toEqual(ActionType.EXPORT_GLOSSARY);
-        expect(successAction.status).toEqual(AsyncActionStatus.FAILURE);
       });
     });
   });
